pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tor-forensic-collector"
version = "1.0.0"
description = "Focused digital forensics tool for extracting and correlating Tor Browser artifacts from Windows systems"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Danilo Solcia", email = "solcia.danilo@gmail.com"}
]
keywords = [
    "forensics",
    "digital-forensics",
    "tor-browser",
    "windows-forensics",
    "incident-response",
    "nfi",
    "artifact-extraction"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Forensics"
]

dependencies = [
    # Core dependencies (always required)
    "typing-extensions>=4.8.0",

    # Registry parsing (NTUSER.DAT)
    "yarp>=1.0.0",  # Pure-Python registry parser

    # EVTX parsing (Windows Event Logs)
    "python-evtx>=0.7.4",  # Pure-Python EVTX parser

    # Prefetch parsing (Windows Prefetch files)
    "windowsprefetch>=4.0.3",  # Prefetch file parser


]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pylint>=2.17.0",
    "flake8>=6.1.0"
]

docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0"
]

all = [
    "tor-forensic-collector[dev,docs]"
]

[project.urls]
"Homepage" = "https://github.com/danisolci/tor-forensic-collector"
"Bug Reports" = "https://github.com/danisolci/tor-forensic-collector/issues"
"Source" = "https://github.com/danisolci/tor-forensic-collector"

[project.scripts]
tor-forensic-collector = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
src = ["py.typed"]

# ==============================================================================
# DEVELOPMENT TOOLS CONFIGURATION
# ==============================================================================

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "yarp",
    "Evtx.*",
    "windowsprefetch"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml"
]

[tool.pylint.main]
fail-under = 8.0
ignore = ["CVS", ".git", "__pycache__", ".pytest_cache"]
jobs = 0
limit-inference-results = 100
py-version = "3.9"

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring (we use comprehensive docstrings, but not everywhere)
    "R0903",  # too-few-public-methods (data classes are OK)
    "R0913",  # too-many-arguments (some forensic functions need many params)
    "W0212",  # protected-access (needed for internal utilities)
]

[tool.pylint.format]
max-line-length = 100

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod"
]

# ==============================================================================
# NOTES ON DEPENDENCIES
# ==============================================================================

# yarp (Yet Another Registry Parser)
# - Pure-Python implementation (no C dependencies)
# - Supports NTUSER.DAT and SYSTEM hives
# - Handles ROT13-encoded UserAssist keys
# - GitHub: https://github.com/msuhanov/yarp

# python-evtx
# - Pure-Python EVTX parser (no dependencies on Windows APIs)
# - Parses Event ID 4688 (Process Creation)
# - Works on Linux for offline analysis
# - GitHub: https://github.com/williballenthin/python-evtx

# windowsprefetch
# - Parses Windows Prefetch files (.pf)
# - Extracts execution timestamps and run counts
# - Supports Windows 7/8/10 formats
# - GitHub: https://github.com/PoorBillionaire/Windows-Prefetch-Parser

# Note: sqlite3 is built-in to Python, no external dependency needed
